remove example utilities
