+++ examples/arkode/C_serial/ark_dissipated_exp_entropy.c
@@ -31,14 +31,37 @@
 #include <math.h>
 #include <stdio.h>
 
-/* Common utilities */
-#include <example_utilities.h>
-
 /* SUNDIALS headers */
 #include <arkode/arkode_arkstep.h>
 #include <nvector/nvector_serial.h>
 #include <sunlinsol/sunlinsol_dense.h>
 #include <sunmatrix/sunmatrix_dense.h>
+
+/* Convince macros for calling precision-specific math functions */
+#if defined(SUNDIALS_DOUBLE_PRECISION)
+#define EXP(x)  (exp((x)))
+#define SQRT(x) (sqrt((x)))
+#define LOG(x)  (log((x)))
+#elif defined(SUNDIALS_SINGLE_PRECISION)
+#define EXP(x)  (expf((x)))
+#define SQRT(x) (sqrtf((x)))
+#define LOG(x)  (logf((x)))
+#elif defined(SUNDIALS_EXTENDED_PRECISION)
+#define EXP(x)  (expl((x)))
+#define SQRT(x) (sqrtl((x)))
+#define LOG(x)  (logl((x)))
+#endif
+
+/* Convince macros for using precision-specific format specifiers */
+#if defined(SUNDIALS_EXTENDED_PRECISION)
+#define GSYM "Lg"
+#define ESYM "Le"
+#define FSYM "Lf"
+#else
+#define GSYM "g"
+#define ESYM "e"
+#define FSYM "f"
+#endif
 
 /* ----------------------- *
  * User-supplied functions *
@@ -63,6 +86,12 @@ int JacEnt(N_Vector y, N_Vector J, void* user_data);
 
 /* Analytic solution */
 int ans(sunrealtype t, N_Vector y);
+
+/* Check for an unrecoverable (negative) return flag from a SUNDIALS function */
+int check_flag(int flag, const char* funcname);
+
+/* Check if a function returned a NULL pointer */
+int check_ptr(void* ptr, const char* funcname);
 
 /* ------------ *
  * Main Program *
@@ -419,5 +448,27 @@ int ans(sunrealtype t, N_Vector y)
 {
   sunrealtype* ydata = N_VGetArrayPointer(y);
   ydata[0] = LOG(EXP(SUN_RCONST(-0.5)) + t);
+  return 0;
+}
+
+/* Check for an unrecoverable (negative) return flag from a SUNDIALS function */
+int check_flag(int flag, const char* funcname)
+{
+  if (flag < 0)
+  {
+    fprintf(stderr, "ERROR: %s() returned %d\n", funcname, flag);
+    return 1;
+  }
+  return 0;
+}
+
+/* Check if a function returned a NULL pointer */
+int check_ptr(void* ptr, const char* funcname)
+{
+  if (!ptr)
+  {
+    fprintf(stderr, "ERROR: %s() returned NULL\n", funcname);
+    return 1;
+  }
   return 0;
 }
